sine wave color but blocky

+++ fractol.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   fractol.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Kai <kdrumm@student.42.us.org>             +#+  +:+       +#+        */
+/*   By: kaidrumm <kaidrumm@student.42.us>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/21 22:50:51 by kdrumm            #+#    #+#             */
-/*   Updated: 2017/03/28 20:04:39 by KaiDrumm         ###   ########.us       */
+/*   Updated: 2017/03/28 20:12:25 by kaidrumm         ###   ########.us       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -65,7 +65,7 @@ void		mandelbrot(t_map *map);
 int			expose_hook(t_map *map);
 int			rgbtoi(t_triple *rgb);
 void		mandelbrot(t_map *map);
-t_triple	*color(int i);
+t_triple	*color(double i);
 void		julia(t_map *map, double cr, double ci);
 void		cubic_mandelbrot(t_map *map);
 void		cubic_julia(t_map *map, double cr, double ci);
